Split implementation of running an artemis node from sending deposits.
