+++ acceptance-tests/src/test-support/java/tech/pegasys/artemis/test/acceptance/dsl/ArtemisDepositSender.java
@@ -1,32 +1,33 @@
+/*
+ * Copyright 2019 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
 package tech.pegasys.artemis.test.acceptance.dsl;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.time.Duration;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.testcontainers.containers.Network;
-import org.testcontainers.containers.wait.strategy.WaitStrategy;
-import org.testcontainers.containers.wait.strategy.WaitStrategyTarget;
 import tech.pegasys.artemis.util.Waiter;
 
-public class ArtemisDepositSender extends AbstractArtemisNode {
-
-  private static final String VALIDATORS_YAML_PATH = "/validators.yaml";
+public class ArtemisDepositSender extends Node {
+  private static final Logger LOG = LogManager.getLogger();
 
   public ArtemisDepositSender(final Network network) {
-    super(network);
+    super(network, ArtemisNode.ARTEMIS_DOCKER_IMAGE, LOG);
   }
 
   public void sendValidatorDeposits(final BesuNode eth1Node, final int numberOfValidators) {
-    container.setWaitStrategy(
-        new WaitStrategy() {
-          @Override
-          public void waitUntilReady(final WaitStrategyTarget waitStrategyTarget) {}
-
-          @Override
-          public WaitStrategy withStartupTimeout(final Duration startupTimeout) {
-            return this;
-          }
-        });
     container.setCommand(
         "validator",
         "generate",
@@ -40,7 +41,6 @@ public class ArtemisDepositSender extends AbstractArtemisNode {
         eth1Node.getInternalJsonRpcUrl());
     container.start();
     Waiter.waitFor(() -> assertThat(container.isRunning()).isFalse());
-    System.out.println(container.getLogs());
     container.stop();
   }
 }
