+++ acceptance-tests/src/test-support/java/tech/pegasys/artemis/test/acceptance/dsl/BesuNode.java
@@ -15,7 +15,6 @@ package tech.pegasys.artemis.test.acceptance.dsl;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.containers.Network;
 import org.testcontainers.containers.wait.strategy.HttpWaitStrategy;
 import org.testcontainers.utility.MountableFile;
@@ -24,39 +23,29 @@ public class BesuNode extends Node {
   private static final Logger LOG = LogManager.getLogger();
   private static final int JSON_RPC_PORT = 8545;
 
-  private final GenericContainer<?> container;
-
   public BesuNode(final Network network) {
-    container =
-        new GenericContainer<>("hyperledger/besu:1.3.6")
-            .withNetwork(network)
-            .withNetworkAliases(nodeAlias)
-            .withExposedPorts(JSON_RPC_PORT)
-            .withLogConsumer(frame -> LOG.debug(frame.getUtf8String().trim()))
-            .waitingFor(new HttpWaitStrategy().forPort(JSON_RPC_PORT).forPath("/liveness"))
-            .withCopyFileToContainer(
-                MountableFile.forClasspathResource("besu/depositContractGenesis.json"),
-                "/genesis.json")
-            .withCommand(
-                "--rpc-http-enabled",
-                "--rpc-http-port",
-                Integer.toString(JSON_RPC_PORT),
-                "--rpc-http-cors-origins=*",
-                "--host-whitelist=*",
-                "--miner-enabled",
-                "--miner-coinbase",
-                "0xfe3b557e8fb62b89f4916b721be55ceb828dbd73",
-                "--genesis-file",
-                "/genesis.json");
+    super(network, "hyperledger/besu:1.3.6", LOG);
+    container
+        .withExposedPorts(JSON_RPC_PORT)
+        .withLogConsumer(frame -> LOG.debug(frame.getUtf8String().trim()))
+        .waitingFor(new HttpWaitStrategy().forPort(JSON_RPC_PORT).forPath("/liveness"))
+        .withCopyFileToContainer(
+            MountableFile.forClasspathResource("besu/depositContractGenesis.json"), "/genesis.json")
+        .withCommand(
+            "--rpc-http-enabled",
+            "--rpc-http-port",
+            Integer.toString(JSON_RPC_PORT),
+            "--rpc-http-cors-origins=*",
+            "--host-whitelist=*",
+            "--miner-enabled",
+            "--miner-coinbase",
+            "0xfe3b557e8fb62b89f4916b721be55ceb828dbd73",
+            "--genesis-file",
+            "/genesis.json");
   }
 
   public void start() {
     container.start();
-  }
-
-  @Override
-  public void stop() {
-    container.stop();
   }
 
   public String getDepositContractAddress() {
